+++ assets/javascripts/kitten/components/buttons/social-button-icon.js
@@ -8,13 +8,13 @@ import { YoutubeIcon } from 'kitten/components/icons/youtube-icon'
 import { PinterestIcon } from 'kitten/components/icons/pinterest-icon'
 
 export const FacebookButtonIcon = props => (
-  <ButtonIcon {...props} modifier="facebook" aria-label="Facebook">
+  <ButtonIcon {...props} modifier="facebook">
     <FacebookIcon className="k-ButtonIcon__svg" />
   </ButtonIcon>
 )
 
 export const TwitterButtonIcon = props => (
-  <ButtonIcon {...props} modifier="twitter" aria-label="Twitter">
+  <ButtonIcon {...props} modifier="twitter">
     <TwitterIcon className="k-ButtonIcon__svg" />
   </ButtonIcon>
 )
@@ -26,19 +26,19 @@ export const LinkedinButtonIcon = props => (
 )
 
 export const InstagramButtonIcon = props => (
-  <ButtonIcon {...props} modifier="instagram" aria-label="Instagram">
+  <ButtonIcon {...props} modifier="instagram">
     <InstagramIcon className="k-ButtonIcon__svg" />
   </ButtonIcon>
 )
 
 export const YoutubeButtonIcon = props => (
-  <ButtonIcon {...props} modifier="youtube" aria-label="Youtube">
+  <ButtonIcon {...props} modifier="youtube">
     <YoutubeIcon className="k-ButtonIcon__svg" />
   </ButtonIcon>
 )
 
 export const PinterestButtonIcon = props => (
-  <ButtonIcon {...props} modifier="pinterest" aria-label="Pinterest">
+  <ButtonIcon {...props} modifier="pinterest">
     <PinterestIcon className="k-ButtonIcon__svg" />
   </ButtonIcon>
 )